Dialog zum Abrufen gespeicherter Kontakte hinzugefügt Legende Testweise hinzugefügt

+++ app/src/main/java/min3d/objectPrimitives/HollowCylinder.java
@@ -85,7 +85,7 @@ public class HollowCylinder extends Object3dContainer {
 
     //
     private void addVerticalSurface(boolean $isOuter) {
-        int off = (int) (_vertices.size() / 2);
+        int off = _vertices.size() / 2;
 
         for (int i = 0; i < _segs - 1; i++) {
             int ul = i * 2;
