+++ app/src/main/java/min3d/parser/ObjParser.java
@@ -252,11 +252,11 @@ public class ObjParser extends AParser implements IParser {
                     subParts = new StringTokenizer(parts.nextToken(), "/");
 
                 int index = i - 1;
-                v[index] = (int) (Integer.parseInt(subParts.nextToken()) - 1);
+                v[index] = Integer.parseInt(subParts.nextToken()) - 1;
                 if (hasuv)
-                    uv[index] = (int) (Integer.parseInt(subParts.nextToken()) - 1);
+                    uv[index] = Integer.parseInt(subParts.nextToken()) - 1;
                 if (hasn)
-                    n[index] = (int) (Integer.parseInt(subParts.nextToken()) - 1);
+                    n[index] = Integer.parseInt(subParts.nextToken()) - 1;
             }
         }
     }
